+++ src/audio_device_module.cpp
@@ -1,5 +1,7 @@
 #include "sora/audio_device_module.h"
 
+#include <api/audio/create_audio_device_module.h>
+
 #if defined(SORA_CPP_SDK_WINDOWS)
 #include <modules/audio_device/include/audio_device_factory.h>
 #endif
@@ -14,19 +16,18 @@ namespace sora {
 webrtc::scoped_refptr<webrtc::AudioDeviceModule> CreateAudioDeviceModule(
     const AudioDeviceModuleConfig& config) {
   if (config.audio_layer == webrtc::AudioDeviceModule::kDummyAudio) {
-    return webrtc::AudioDeviceModule::Create(
-        webrtc::AudioDeviceModule::kDummyAudio, config.task_queue_factory);
+    return webrtc::CreateAudioDeviceModule(
+        config.env, webrtc::AudioDeviceModule::kDummyAudio);
   }
 
 #if defined(SORA_CPP_SDK_WINDOWS)
   return webrtc::CreateWindowsCoreAudioAudioDeviceModule(
-      config.task_queue_factory);
+      &config.env.task_queue_factory());
 #elif defined(SORA_CPP_SDK_ANDROID)
-  return webrtc::CreateOpenSLESAudioDeviceModule(
-      (JNIEnv*)config.jni_env, (jobject)config.application_context);
+  return webrtc::CreateJavaAudioDeviceModule(
+      (JNIEnv*)config.jni_env, config.env, (jobject)config.application_context);
 #else
-  return webrtc::AudioDeviceModule::Create(config.audio_layer,
-                                           config.task_queue_factory);
+  return webrtc::CreateAudioDeviceModule(config.env, config.audio_layer);
 #endif
 }