+++ CHANGES.md
@@ -23,11 +23,13 @@
   - @melpon
 - [ADD] インストールするファイルに cmake/android.toolchain.cmake を追加
   - @melpon
-- [UPDATE] libwebrtc を m137.7151.0.0 にあげる
+- [UPDATE] libwebrtc を m138.7204.0.0 にあげる
   - `rtc::revive::` を　`webrtc::revive::` に変更する
   - m137 で `rtc_base/third_party/base64/base64.h` が削除されたため、websocket.cpp で `rtc_base/base64.h` をインクルードするように変更する
   - `webrtc::Base64::Encode` から `webrtc::Base64Encode` に変更する
   - Android ビルドに利用するコンパイラを Android NDK に内包されている clang ではなく、libwebrtc の clang に変更する
+  - `sora::AudioDeviceModuleConfig` から `task_queue_factory` を削除して `env` を追加する
+  - `webrtc::CreateOpenSLESAudioDeviceModule` を `webrtc::CreateJavaAudioDeviceModule` に変更する
   - @miosakuma @torikizi @melpon
 - [UPDATE] Android NDK を r28b にあげる
   - 16KB ページサイズに対応するため